axon-bulk-export/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
ez-setup==0.9
idna==3.3
lxml==4.8.0
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
